refactor(lightning): move LND and c-lightning libs into lightning folder
