+++ src/lib/lightning/clightning/types.ts
