+++ src/lib/lightning/lnd/lndProxyClient.spec.ts
