+++ src/lib/lightning/lnd/lndProxyClient.ts
