+++ src/lib/lightning/lnd/lndService.spec.ts
